+++ packages/dartway_serverpod_core/dartway_serverpod_core_server/lib/src/domain/crud/model_configs/dw_get_model_list_config.dart
@@ -10,6 +10,7 @@ class DwGetModelListConfig<Model extends TableRow>
     extends DwGetConfigInterface<Model> implements DwGetListInterface<Model> {
   const DwGetModelListConfig({
     required super.accessFilter,
+    super.allowAnonymous,
     super.include,
     // this.additionalModelsFetchFunction,
     super.defaultOrderByList,